requirements.txt
APScheduler==3.5.1
Django==2.0
gunicorn
pycryptodome==3.6.6
djangorestframework==3.7.7
django-cors-headers==2.1.0
requests==2.18.4
django-filter==1.1.0
psycopg2==2.7.4
django-rest-swagger==2.2.0
coverage==4.5.2
django-nose==1.4.6
jsonnet==0.12.1
django-heroku==0.3.1
pyexcel-xlsx
django_excel
httplib2==0.15.0
Markdown==3.1.1